[audio] fuchsia_audio_device::Registry
This CL implements the fuchsia_audio_device::Registry
protocol, and adds support in the Device class for
on-the-fly device removal.

Bug: 87650

Multiply: audio_services_device_registry_unittests
Multiply: audio_services_device_registry_warning_unittests

Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/730030
Commit-Queue: Martin Puryear <mpuryear@google.com>

+++ src/media/audio/services/device_registry/provider_server.h
@@ -33,14 +33,14 @@ class ProviderServer
   void AddDevice(AddDeviceRequest& request, AddDeviceCompleter::Sync& completer) override;
 
   // Static object count, for debugging purposes.
-  static uint32_t count() { return count_; }
+  static uint64_t count() { return count_; }
 
  private:
   template <typename ServerT, template <typename T> typename FidlServerT, typename ProtocolT>
   friend class BaseFidlServer;
 
   static inline const std::string_view kClassName = "ProviderServer";
-  static uint32_t count_;
+  static uint64_t count_;
 
   explicit ProviderServer(std::shared_ptr<AudioDeviceRegistry> parent);
